Add remove bank works. Some issues with webview.
