+++ config.xml
@@ -3,7 +3,7 @@
     <name>Northbricks</name>
     <description>http://www.northbricks.com</description>
     <author email="info@northbricks.com" href="http://www.northbricks.io/">Northbricks Team</author>
-    <content src="index.html" />
+    <content original-src="index.html" src="http://192.168.51.166:8100" />
     <access origin="*" />
     <allow-navigation href="http://ionic.local/*" />
     <allow-intent href="http://*/*" />
@@ -92,12 +92,15 @@
     <plugin name="cordova-plugin-actionsheet" spec="^2.3.3" />
     <plugin name="cordova-plugin-statusbar" spec="git+https://github.com/apache/cordova-plugin-statusbar.git" />
     <allow-navigation href="http://192.168.8.108:8100" />
-    <plugin name="cordova-plugin-inappbrowser" spec="^2.0.2" />
     <allow-navigation href="http://10.177.49.55:8100" />
     <allow-navigation href="http://10.117.3.184:8100" />
     <allow-navigation href="http://192.168.108.187:8101" />
     <allow-navigation href="http://172.20.10.2:8101" />
     <allow-navigation href="http://172.22.94.161:8100" />
-    <engine name="ios" spec="4.5.4" />
-    <engine name="android" spec="7.1.0" />
+    <allow-navigation href="http://192.168.1.43:8101" />
+    <allow-navigation href="http://192.168.1.43:8100" />
+    <plugin name="cordova-plugin-inappbrowser" spec="^2.0.2" />
+    <allow-navigation href="http://192.168.51.83:8100" />
+    <allow-navigation href="http://192.168.51.166:8100" />
+    <engine name="ios" spec="^4.5.4" />
 </widget>
