+++ src/pages/profile/profile.html
@@ -31,12 +31,12 @@
     <ion-list-header>
       My banks
     </ion-list-header>
-    <ion-item *ngFor="let bank of banks" (click)="showBank(bank)">
+    <ion-item *ngFor="let bank of banks" (click)="removeBank(bank.id)">
       <ion-avatar item-start class="logo">
         <img src="../../assets/logo/Nordea200px.png">
       </ion-avatar>
       <h2>{{bank.fullName}}</h2>
-      <button ion-button outline item-end (click)="removeBank()">Disconnect</button>
+      <button ion-button outline item-end>Disconnect</button>
     </ion-item>
   </ion-list>
   <!-- <ion-list>
